pyproject.toml
[project]
name = "learndl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "baostock>=0.8.9",
    "catboost>=1.2.8",
    "colorlog>=6.9.0",
    "cvxopt>=1.3.2",
    "cvxpy>=1.7.2",
    "dask>=2025.7.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "lightgbm>=4.6.0",
    "matplotlib==3.10.3",
    "mosek>=11.0.28",
    "numpy==1.26.0",
    "openpyxl>=3.1.5",
    "optuna>=4.5.0",
    "pandas==2.3.0",
    "pip>=25.2",
    "plottable>=0.1.5",
    "polars>=1.32.3",
    "portalocker>=3.2.0",
    "pymysql>=1.1.2",
    "pypinyin>=0.55.0",
    "pyreadr>=0.5.3",
    "pytimedinput>=2.0.1",
    "pyyaml>=6.0.2",
    "rqdatac>=3.2.11",
    "ruff>=0.13.0",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "streamlit==1.46.0",
    "streamlit-ace>=0.1.1",
    "streamlit-autorefresh>=1.0.1",
    "tensorboard>=2.20.0",
    "torch==2.7.1",
    "tushare>=1.4.24",
    "xarray>=2025.8.0",
    "xgboost>=3.0.4",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "resources"
]

[tool.ruff.lint]
# 2. Avoid enforcing line-length violations (`E501`)
ignore = [
    "E401" , # Multiple imports on one line
    "F541" , # f-string without any placeholders
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401" , # unused imports
    "F403" , # import *
]
"required_pkgs.py" = ["F401"]
